Swapped out the entire file

+++ database/migrations/2024_01_01_000008_create_taskable_table.php
@@ -8,15 +8,17 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('taskables', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('task_id');
-            $table->morphs('taskable');
-            $table->timestamps();
+        if (!Schema::hasTable('taskables')) {
+            Schema::create('taskables', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('task_id');
+                $table->morphs('taskable');
+                $table->timestamps();
 
-            $table->foreign('task_id')->references('id')->on('tasks')->onDelete('cascade');
-            $table->unique(['task_id', 'taskable_id', 'taskable_type']);
-        });
+                $table->foreign('task_id')->references('id')->on('tasks')->onDelete('cascade');
+                $table->unique(['task_id', 'taskable_id', 'taskable_type']);
+            });
+        }
     }
 
     public function down(): void
